requirements.txt
psycopg2-binary==2.9.6
Flask==2.3.2
requests==2.31.0
gunicorn==20.1.0
